Fix some bugs
Hello! Thanks for the useful tool.

I fixed a bug in the code that builds the rows / columns (lines 69-70 need the value converted to an int). 

I also promoted the master zdefocus node to a tab so the user has access to it without going inside the gizmo. 

I also disabled the copy and remove nodes on the filter input that was causing the node to error in all cases unless the filter input had an alpha channel. This lets the user use the default "bladed" or "disc" settings on the zdefocus if they want.

Hope this helps! It could probably be better still!